ADD: 두가지 필터링 조건 정렬
